created skeleton for logging facilities on server
Created a loggin.properties file in logs/logging.properties. This file specifies where logs are saved (log/messages.log.0). This properties file is loaded in StartProgram.java. If the properties file isn't found server prints and error to stderr and exits.

Added a messageLogger instance variable in all relevant classes. This variable is used to log to the log file. Logging in all classes will be saved to the same file.